makesd: add new -x option to install files and packages
The problem with current -b -k -m -c -d options is that the
list of files is hard-coded into makesd code which is not very
flexible.

Many boards (and their SD cards) need different file names
for u-boot and kernel image.

The new -x option introduces a lower-level mechanism where
the caller (or a macro) can specify a file on some server (URI)
and a local file name or local directory. If a file is given
it is tried to be loaded and renamed. If a directory is given
it is by default installed in this directory. If a file is not
found, but one with .tgz or .tbz appended, it is unpacked
into the given directory.

Multiple -x specifiers are collected (for each partition) and
applied one after the other.

With that, it should be possible to reimplement the commands
-b -k -m -c -d as macros which make use of the basic -x
mechanism. This step is not yet included in this patch because
we need to gain more experience with it first.

Signed-off-by: H. Nikolaus Schaller <hns@goldelico.com>