Revert "our letux fstab may mount multiple boot partitions to /boot and hide the one we want to reflash"
This reverts commit f1af8f61d2e60ea08b74e801e4dec6ceef1764b0.

Problem reported:

I don't think it should be unmounting the /boot partition on the machine running the script there.
It caused some issues booting here, since I had upgraded the kernel after running makesd, but it didn't end up on my /boot partition...

I think it is this:

unter certain conditions there may be two mounts active on /boot, one hiding the other.
If the hidden one is the eMMC boot partition and the Pyra has been booted from SD card,
it is not possible to unmount the eMMC boot partition (because it is hidden).

The idea was to actively umount /boot (twice) to make sure it can be mounted as /media/P1.
This appears to be only be needed if one runs makesd on the Pyra to flash eMMC that way
and has a fstab with multiple auto entries for /boot.

I am not aware that another device needs this. Maybe the BeagleBoneBlack.

So we can revert.

Reported-By: atc@k-n-p.org